refactor: break upward edges before presentation regroup (#1087)
Prepares the presentation move by removing the domain -> presentation edges
that would otherwise appear once the files relocate.

- Extract BuildAgentMessagesFromEntries out of services/chatcompletion into
  internal/conversation, so internal/agent and cmd/headless stop importing a
  package that is about to become TUI code.
- Split channels/contracts.go into a new internal/channels port package and
  move the Telegram driver, channel manager and channel commands to
  internal/presentation/telegram. internal/scheduler now depends on the port,
  not on the driver.
- Move prune.go to internal/platform/utils and export PruneFilesByModTime; it
  is shared by frame sources, the TUI and the Telegram session image dir.
