+++ internal/presentation/telegram/retention_test.go
@@ -1,4 +1,4 @@
-package channels
+package telegram
 
 import (
 	"fmt"
