+++ internal/presentation/telegram/telegram.go
@@ -1,4 +1,4 @@
-package channels
+package telegram
 
 import (
 	"bytes"
@@ -28,6 +28,7 @@ import (
 
 	config "github.com/inference-gateway/cli/config"
 	agentdomain "github.com/inference-gateway/cli/internal/agent/domain"
+	channels "github.com/inference-gateway/cli/internal/channels"
 	logger "github.com/inference-gateway/cli/internal/platform/logger"
 )
 
@@ -70,7 +71,7 @@ type VoiceTranscriber interface {
 	TranscribeFile(ctx context.Context, audioPath string) (string, error)
 }
 
-// TelegramChannel implements Channel for the Telegram Bot API
+// TelegramChannel implements channels.Channel for the Telegram Bot API
 // using the go-telegram/bot SDK with long-polling.
 type TelegramChannel struct {
 	cfg         config.TelegramChannelConfig
@@ -83,7 +84,7 @@ type TelegramChannel struct {
 	media *FileRetention
 
 	// commands to advertise via SetMyCommands on Start.
-	commands []ChannelCommand
+	commands []channels.ChannelCommand
 
 	// msgIDs tracks recent message IDs per chat so ClearHistory can delete them.
 	// ponytail: in-memory + capped; Telegram can't delete >48h-old messages anyway, persistence buys nothing
@@ -108,7 +109,7 @@ func NewTelegramChannel(cfg config.TelegramChannelConfig, transcriber VoiceTrans
 }
 
 // SetCommands sets the slash commands advertised to Telegram on Start.
-func (t *TelegramChannel) SetCommands(cmds []ChannelCommand) {
+func (t *TelegramChannel) SetCommands(cmds []channels.ChannelCommand) {
 	t.commands = cmds
 }
 
@@ -124,7 +125,7 @@ func (t *TelegramChannel) trackMessage(chatID int64, messageID int) {
 }
 
 // ClearHistory best-effort deletes the tracked messages of a chat.
-// Implements HistoryCleaner. Telegram silently skips messages it can no
+// Implements channels.HistoryCleaner. Telegram silently skips messages it can no
 // longer delete (older than 48h), and batches that fail are logged and skipped.
 func (t *TelegramChannel) ClearHistory(ctx context.Context, recipientID string) error {
 	if t.bot == nil {
@@ -154,7 +155,7 @@ func (t *TelegramChannel) ClearHistory(ctx context.Context, recipientID string) 
 
 // telegramBotCommands converts channel commands to Telegram bot commands,
 // dropping names Telegram would reject and truncating long descriptions.
-func telegramBotCommands(cmds []ChannelCommand) []models.BotCommand {
+func telegramBotCommands(cmds []channels.ChannelCommand) []models.BotCommand {
 	var out []models.BotCommand
 	for _, c := range cmds {
 		if !botCommandNameRe.MatchString(c.Name) {
@@ -172,7 +173,7 @@ func telegramBotCommands(cmds []ChannelCommand) []models.BotCommand {
 // inlineKeyboard renders message buttons as a one-button-per-row inline
 // keyboard (full-width, thumb-friendly). Returns nil for no buttons so callers
 // can pass it straight to sendText.
-func inlineKeyboard(buttons []MessageButton) models.ReplyMarkup {
+func inlineKeyboard(buttons []channels.MessageButton) models.ReplyMarkup {
 	if len(buttons) == 0 {
 		return nil
 	}
@@ -193,7 +194,7 @@ func (t *TelegramChannel) Name() string {
 }
 
 // Start begins long-polling for Telegram updates and sends inbound messages to the inbox
-func (t *TelegramChannel) Start(ctx context.Context, inbox chan<- InboundMessage) error {
+func (t *TelegramChannel) Start(ctx context.Context, inbox chan<- channels.InboundMessage) error {
 	if t.cfg.BotToken == "" {
 		return fmt.Errorf("telegram bot token is required")
 	}
@@ -284,7 +285,7 @@ func (t *TelegramChannel) Start(ctx context.Context, inbox chan<- InboundMessage
 }
 
 // Send delivers a message through the Telegram Bot API
-func (t *TelegramChannel) Send(ctx context.Context, msg OutboundMessage) error {
+func (t *TelegramChannel) Send(ctx context.Context, msg channels.OutboundMessage) error {
 	if t.bot == nil {
 		return fmt.Errorf("telegram bot not started")
 	}
@@ -333,7 +334,7 @@ func (t *TelegramChannel) sendImageFile(ctx context.Context, chatID int64, p str
 
 	if err != nil {
 		logger.Warn("sendPhoto failed, retrying as document", "path", p, "error", err)
-		f2, oerr := os.Open(p) // SendPhoto consumed the first reader; reopen for the retry.
+		f2, oerr := os.Open(p)
 		if oerr != nil {
 			logger.Warn("skipping outbound image", "path", p, "error", oerr)
 			return
@@ -380,7 +381,7 @@ func (t *TelegramChannel) sendText(ctx context.Context, chatID int64, text strin
 }
 
 // SendApproval sends a tool approval prompt with inline keyboard buttons.
-// Implements ApprovalChannel.
+// Implements channels.ApprovalChannel.
 func (t *TelegramChannel) SendApproval(ctx context.Context, recipientID string, req *ipc.ApprovalRequest) error {
 	if t.bot == nil {
 		return fmt.Errorf("telegram bot not started")
@@ -410,10 +411,10 @@ func (t *TelegramChannel) Stop() error {
 	return nil
 }
 
-// processUpdate converts a Telegram update into an InboundMessage (or nil if
+// processUpdate converts a Telegram update into an channels.InboundMessage (or nil if
 // skipped). Videos are only accepted when mediaEnabled (channels.telegram.media
 // saving is on); otherwise they are skipped as before.
-func processUpdate(update *models.Update, mediaEnabled bool) *InboundMessage {
+func processUpdate(update *models.Update, mediaEnabled bool) *channels.InboundMessage {
 	if update.CallbackQuery != nil {
 		return processCallbackQuery(update.CallbackQuery)
 	}
@@ -449,7 +450,7 @@ func processUpdate(update *models.Update, mediaEnabled bool) *InboundMessage {
 		metadata["user_id"] = strconv.FormatInt(msg.From.ID, 10)
 	}
 
-	inbound := &InboundMessage{
+	inbound := &channels.InboundMessage{
 		ChannelName: "telegram",
 		SenderID:    senderID,
 		Content:     content,
@@ -497,7 +498,7 @@ func processUpdate(update *models.Update, mediaEnabled bool) *InboundMessage {
 // returns false when the message should be dropped (transcription disabled,
 // failed, or produced no text), matching the prior behavior of ignoring voice
 // messages the bot cannot turn into text.
-func (t *TelegramChannel) applyVoiceTranscription(ctx context.Context, b *bot.Bot, msg *InboundMessage, fileID string) bool {
+func (t *TelegramChannel) applyVoiceTranscription(ctx context.Context, b *bot.Bot, msg *channels.InboundMessage, fileID string) bool {
 	if t.transcriber == nil {
 		logger.Warn("received a voice message but speech-to-text is disabled; ignoring")
 		return false
@@ -575,7 +576,7 @@ func (t *TelegramChannel) mediaAcceptable(mime string, size int64) error {
 // directory after size/mime checks, then records the saved path in the message
 // metadata and content so the agent can use the file as an asset. Best-effort:
 // a rejected or failed save never drops the message.
-func (t *TelegramChannel) saveInboundMedia(msg *InboundMessage, mime, hintPath string, data []byte) {
+func (t *TelegramChannel) saveInboundMedia(msg *channels.InboundMessage, mime, hintPath string, data []byte) {
 	if t.media == nil {
 		return
 	}
@@ -596,7 +597,7 @@ func (t *TelegramChannel) saveInboundMedia(msg *InboundMessage, mime, hintPath s
 // Telegram declared (so oversized files are rejected before any download),
 // then fetches the video and saves it via saveInboundMedia. Rejections are
 // appended to the message content so the agent can tell the user why.
-func (t *TelegramChannel) downloadInboundVideo(ctx context.Context, b *bot.Bot, msg *InboundMessage, fileID string) {
+func (t *TelegramChannel) downloadInboundVideo(ctx context.Context, b *bot.Bot, msg *channels.InboundMessage, fileID string) {
 	mime := msg.Metadata["media_mime"]
 	size, _ := strconv.ParseInt(msg.Metadata["media_size"], 10, 64)
 	if err := t.mediaAcceptable(mime, size); err != nil {
@@ -666,16 +667,16 @@ func mimeFromPath(path string) string {
 }
 
 // processCallbackQuery converts a Telegram callback query (e.g., inline button click)
-// into an InboundMessage. Approval buttons carry approval metadata; command
+// into an channels.InboundMessage. Approval buttons carry approval metadata; command
 // buttons ("/..." callback data) are delivered as if the user typed the
 // command. Returns nil for unrecognized data.
-func processCallbackQuery(cq *models.CallbackQuery) *InboundMessage {
+func processCallbackQuery(cq *models.CallbackQuery) *channels.InboundMessage {
 	if strings.HasPrefix(cq.Data, "/") {
 		chatID := callbackChatID(cq)
 		if chatID == 0 {
 			return nil
 		}
-		return &InboundMessage{
+		return &channels.InboundMessage{
 			ChannelName: "telegram",
 			SenderID:    strconv.FormatInt(chatID, 10),
 			Content:     cq.Data,
@@ -706,7 +707,7 @@ func processCallbackQuery(cq *models.CallbackQuery) *InboundMessage {
 		return nil
 	}
 
-	return &InboundMessage{
+	return &channels.InboundMessage{
 		ChannelName: "telegram",
 		SenderID:    strconv.FormatInt(chatID, 10),
 		Content:     action,
